refactor: MemberRepository 네이밍, 패키징 변경
